appconfig: Use standard API client factory.
